[compiler] transform fire calls
This is the diff with the meaningful changes. The approach is:
1. Collect fire callees and remove fire() calls, create a new binding for the useFire result
2. Update LoadLocals for captured callees to point to the useFire result
3. Update function context to reference useFire results
4. Insert useFire calls after getting to the component scope

This approach aims to minimize the amount of new bindings we introduce for the function expressions
to minimize bookkeeping for dependency arrays. We keep all of the LoadLocals leading up to function
calls as they are and insert new instructions to load the originally captured function, call useFire,
and store the result in a new promoted temporary. The lvalues that referenced the original callee are
changed to point to the new useFire result.

This is the minimal diff to implement the expected behavior (up to importing the useFire call, next diff)
and further stacked diffs implement error handling. The rules for fire are:
1. If you use fire for a callee in the effect once you must use it for every time you call it in that effect
2. You can only use fire in a useEffect lambda/functions defined inside the useEffect lambda

There is still more work to do here, like updating the effect dependency array and handling object methods

--